backend/requirements.txt
Django==5.2.3
asgiref==3.8.1
django-cors-headers==4.7.0
django-storages==1.14.2
django-reversion==5.1.0
django-environ==0.11.2
django-filter==23.5
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
drf-spectacular==0.27.0
drf-yasg==1.21.7
Pillow==11.2.1
python-decouple==3.8
boto3==1.34.14
botocore==1.34.14

# Integraciones de pago
paypalrestsdk==1.13.1
stripe==7.13.0
requests==2.31.0

# Facturación electrónica
zeep==4.2.1  # Cliente SOAP para servicios web
xmlsec==1.3.13  # Manejo de firmas XML

# Notificaciones por email
django-anymail==10.2  # Integración con proveedores de email

# Testing
pytest==7.4.4
pytest-django==4.7.0
coverage==7.4.1
